@@ -13,31 +13,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^10.1.7",
-    "@angular/animations": "~10.0.3",
-    "@angular/cdk": "^10.1.3",
-    "@angular/cdk-experimental": "^10.1.3",
-    "@angular/common": "~10.0.3",
-    "@angular/compiler": "~10.0.3",
-    "@angular/core": "~10.0.3",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "~10.0.3",
-    "@angular/material": "^10.1.3",
-    "@angular/platform-browser": "~10.0.3",
-    "@angular/platform-browser-dynamic": "~10.0.3",
-    "@angular/router": "~10.0.3",
+    "@angular-devkit/core": "^14.2.10",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/cdk-experimental": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/flex-layout": "^14.0.0-beta.41",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@ckeditor/ckeditor5-adapter-ckfinder": "^23.0.0",
-    "@ckeditor/ckeditor5-angular": "^1.2.3",
+    "@ckeditor/ckeditor5-angular": "^4.0.0",
     "@ckeditor/ckeditor5-autoformat": "^23.0.0",
     "@ckeditor/ckeditor5-basic-styles": "^23.0.0",
     "@ckeditor/ckeditor5-block-quote": "^23.0.0",
-    "@ckeditor/ckeditor5-build-classic": "^22.0.0",
-    "@ckeditor/ckeditor5-build-inline": "^22.0.0",
+    "@ckeditor/ckeditor5-build-classic": "^23.1.0",
+    "@ckeditor/ckeditor5-build-inline": "^23.1.0",
     "@ckeditor/ckeditor5-ckfinder": "^23.0.0",
     "@ckeditor/ckeditor5-dev-utils": "^23.6.1",
     "@ckeditor/ckeditor5-dev-webpack-plugin": "^23.6.1",
     "@ckeditor/ckeditor5-essentials": "^23.0.0",
-    "@ckeditor/ckeditor5-font": "^22.0.0",
+    "@ckeditor/ckeditor5-font": "^23.1.0",
     "@ckeditor/ckeditor5-heading": "^23.0.0",
     "@ckeditor/ckeditor5-image": "^23.0.0",
     "@ckeditor/ckeditor5-indent": "^23.0.0",
@@ -49,47 +49,55 @@
     "@ckeditor/ckeditor5-upload": "^23.0.0",
     "@material/textfield": "^7.0.0",
     "@wiris/mathtype-ckeditor5": "^7.24.1",
-    "angular-gridster2": "^10.1.5",
-    "angular-resize-event": "^2.0.0",
+    "angular-gridster2": "^14.1.2",
+    "angular-resize-event": "^3.2.0",
+    "assert": "^2.0.0",
+    "browserify-fs": "^1.0.0",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^6.0.3",
     "chart.js": "^2.9.3",
     "ckeditor5-custom-build": "file:src/app/ckeditor5",
     "dompurify": "^2.2.0",
-    "jdenticon": "^3.0.1",
+    "jdenticon": "^3.2.0",
     "ng2-charts": "^2.4.1",
-    "ngx-cookie-service": "^10.1.1",
-    "ngx-jdenticon": "^1.0.3",
-    "ngx-tui": "^2.0.0",
+    "ngx-cookie-service": "^14.0.1",
+    "ngx-jdenticon": "^1.0.4",
     "ngx-ui-scroll": "^1.11.0-rc.1",
+    "node-polyfill-webpack-plugin": "^2.0.1",
     "postcss-loader": "^3.0.0",
     "primeicons": "^4.0.0",
+    "process": "^0.11.10",
     "raw-loader": "^3.1.0",
-    "rxjs": "^6.5.5",
+    "rxjs": "^7.8.0",
     "rxjs-observable": "0.0.7",
     "socket.io": "^2.3.0",
+    "stream-browserify": "^3.0.0",
     "style-loader": "^1.2.1",
     "tslib": "^2.0.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.12",
-    "zone.js": "~0.10.3"
+    "util": "^0.12.5",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^5.0.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.2",
-    "@angular/cli": "~10.0.2",
-    "@angular/compiler-cli": "~10.0.3",
-    "@types/jasmine": "~3.5.0",
+    "@angular-builders/custom-webpack": "14.1.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~4.6.4"
   }
 }
