@@ -1,8 +1,8 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ProfileSettingsComponent } from './profile-settings.component';
 import { UserPreferencesService } from '@shared/user-preferences.service';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 import { AppConfigService } from 'src/app/app-config.service';
 
 
@@ -19,10 +19,10 @@ describe('ProfileSettingsComponent', () => {
   let component: ProfileSettingsComponent;
   let fixture: ComponentFixture<ProfileSettingsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       providers:[
-        FormBuilder,
+        UntypedFormBuilder,
         {provide: UserPreferencesService, useClass: UserPreferencesMock},
         {provide: AppConfigService, useValue: { getConfig }}
       ],
