@@ -60,7 +60,7 @@ import {MatSlideToggleModule} from '@angular/material/slide-toggle';
 import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import { LogbookSearchPipe } from './overview/logbook-search.pipe';
 import { NavigationButtonComponent } from './logbook/navigation-button/navigation-button.component'; 
-import { AngularResizedEventModule } from 'angular-resize-event';
+import { AngularResizeEventModule } from 'angular-resize-event';
 import { SnippetContentComponent } from '@shared/snippet/snippet-content/snippet-content.component';
 import {CookieService} from 'ngx-cookie-service';
 import { NgxJdenticonModule, JDENTICON_CONFIG } from 'ngx-jdenticon';
@@ -88,119 +88,118 @@ const appConfigInitializerFn = (appConfig: AppConfigService) => {
 };
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    SnippetComponent,
-    AddContentComponent,
-    SnippetTableComponent,
-    DashboardItemComponent,
-    LogbookItemComponent,
-    DashboardComponent,
-    ChatComponent,
-    TodosComponent,
-    ChartComponent,
-    LoginComponent,
-    OverviewComponent,
-    LogbookComponent,
-    LogbookWidgetComponent,
-    ViewWidgetComponent,
-    ToolbarComponent,
-    WidgetPreferencesComponent,
-    SettingsComponent,
-    CollectionWidgetComponent,
-    AddLogbookComponent,
-    AddCollectionComponent,
-    LogbookSearchPipe,
-    NavigationButtonComponent,
-    SnippetContentComponent,
-    SnippetViewerComponent,
-    SnippetDashboardNameComponent,
-    ViewSettingsComponent,
-    ProfileSettingsComponent,
-    ViewEditComponent,
-    TagEditorComponent,
-    SnippetInfoComponent,
-    SnippetInfoSectionComponent,
-    HotkeysComponent,
-    ExportDialogComponent,
-    DownloadComponent,
-    SearchComponent,
-    SearchWindowComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    MatButtonModule,
-    MatCardModule,
-    MatDialogModule,
-    MatExpansionModule,
-    MatIconModule,
-    MatProgressSpinnerModule,
-    MatSidenavModule,
-    BrowserAnimationsModule,
-    CKEditorModule,
-    HttpClientModule,
-    ScrollingModule,
-    DragDropModule,
-    ExperimentalScrollingModule,
-    MatMenuModule,
-    MatChipsModule,
-    MatFormFieldModule,
-    MatToolbarModule,
-    MatTableModule,
-    MatButtonToggleModule,
-    GridsterModule,
-    MatDividerModule,
-    MatBadgeModule,
-    ReactiveFormsModule,
-    MatInputModule,
-    ChartsModule,
-    MatTooltipModule,
-    MatSelectModule,
-    MatRadioModule,
-    MatCheckboxModule,
-    FormsModule,
-    MatSlideToggleModule,
-    MatAutocompleteModule,
-    AngularResizedEventModule,
-    NgxJdenticonModule,
-    MatTabsModule,
-    UiScrollModule,
-    MatProgressBarModule,
-    MatSnackBarModule
-  ],
-  entryComponents: [HotkeysComponent],
-  providers: [
-    AppConfigService,
-    {
-      provide: APP_INITIALIZER,
-      useFactory: appConfigInitializerFn,
-      multi: true,
-      deps: [AppConfigService],
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthInterceptor,
-      multi: true,
-      deps: [AuthService, Router],
-    },
-    CookieService,
-    { 
-      // Custom identicon style
-      provide: JDENTICON_CONFIG,
-      useValue: {
-        lightness: {
-          color: [0.31, 0.54],
-          grayscale: [0.63, 0.82],
+    declarations: [
+        AppComponent,
+        SnippetComponent,
+        AddContentComponent,
+        SnippetTableComponent,
+        DashboardItemComponent,
+        LogbookItemComponent,
+        DashboardComponent,
+        ChatComponent,
+        TodosComponent,
+        ChartComponent,
+        LoginComponent,
+        OverviewComponent,
+        LogbookComponent,
+        LogbookWidgetComponent,
+        ViewWidgetComponent,
+        ToolbarComponent,
+        WidgetPreferencesComponent,
+        SettingsComponent,
+        CollectionWidgetComponent,
+        AddLogbookComponent,
+        AddCollectionComponent,
+        LogbookSearchPipe,
+        NavigationButtonComponent,
+        SnippetContentComponent,
+        SnippetViewerComponent,
+        SnippetDashboardNameComponent,
+        ViewSettingsComponent,
+        ProfileSettingsComponent,
+        ViewEditComponent,
+        TagEditorComponent,
+        SnippetInfoComponent,
+        SnippetInfoSectionComponent,
+        HotkeysComponent,
+        ExportDialogComponent,
+        DownloadComponent,
+        SearchComponent,
+        SearchWindowComponent,
+    ],
+    imports: [
+        BrowserModule,
+        AppRoutingModule,
+        MatButtonModule,
+        MatCardModule,
+        MatDialogModule,
+        MatExpansionModule,
+        MatIconModule,
+        MatProgressSpinnerModule,
+        MatSidenavModule,
+        BrowserAnimationsModule,
+        CKEditorModule,
+        HttpClientModule,
+        ScrollingModule,
+        DragDropModule,
+        ExperimentalScrollingModule,
+        MatMenuModule,
+        MatChipsModule,
+        MatFormFieldModule,
+        MatToolbarModule,
+        MatTableModule,
+        MatButtonToggleModule,
+        GridsterModule,
+        MatDividerModule,
+        MatBadgeModule,
+        ReactiveFormsModule,
+        MatInputModule,
+        ChartsModule,
+        MatTooltipModule,
+        MatSelectModule,
+        MatRadioModule,
+        MatCheckboxModule,
+        FormsModule,
+        MatSlideToggleModule,
+        MatAutocompleteModule,
+        AngularResizeEventModule,
+        NgxJdenticonModule,
+        MatTabsModule,
+        UiScrollModule,
+        MatProgressBarModule,
+        MatSnackBarModule
+    ],
+    providers: [
+        AppConfigService,
+        {
+            provide: APP_INITIALIZER,
+            useFactory: appConfigInitializerFn,
+            multi: true,
+            deps: [AppConfigService],
         },
-        saturation: {
-          color: 0.50,
-          grayscale: 0.50,
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: AuthInterceptor,
+            multi: true,
+            deps: [AuthService, Router],
         },
-        backColor: '#222',
-      },
-    }
-  ],
-  bootstrap: [AppComponent]
+        CookieService,
+        {
+            // Custom identicon style
+            provide: JDENTICON_CONFIG,
+            useValue: {
+                lightness: {
+                    color: [0.31, 0.54],
+                    grayscale: [0.63, 0.82],
+                },
+                saturation: {
+                    color: 0.50,
+                    grayscale: 0.50,
+                },
+                backColor: '#222',
+            },
+        }
+    ],
+    bootstrap: [AppComponent]
 })
 export class AppModule { }
