@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { HotkeysComponent } from './hotkeys.component';
 import { MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
@@ -7,7 +7,7 @@ describe('HotkeysComponent', () => {
   let component: HotkeysComponent;
   let fixture: ComponentFixture<HotkeysComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       providers: [
         { provide: MAT_DIALOG_DATA, useValue: {} },
