@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { MatDialogModule } from '@angular/material/dialog';
 import { ActivatedRoute, RouterModule } from '@angular/router';
 import {RouterTestingModule} from '@angular/router/testing'
@@ -12,7 +12,7 @@ describe('DashboardItemComponent', () => {
   let viewsSpy:any;
   viewsSpy = jasmine.createSpyObj("ViewsService", ["getLogbookViews"]);
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       providers: [
         {provide: ViewsService, useClass: ViewsServiceMock}
