@@ -1,5 +1,5 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppConfigService } from '../app-config.service';
@@ -12,7 +12,7 @@ describe('LogbookComponent', () => {
   let component: LogbookComponent;
   let fixture: ComponentFixture<LogbookComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ LogbookComponent ],
       imports: [RouterTestingModule, HttpClientTestingModule],
