@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { NavigationButtonComponent } from './navigation-button.component';
 import { ActivatedRoute, ParamMap, Router } from '@angular/router';
@@ -20,7 +20,7 @@ describe('NavigationButtonComponent', () => {
     parent: { url: of(queryParams) },
     snapshot: { queryParams: {id: '1234'}}
  };
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       providers:[
         { provide: ActivatedRoute, useValue: activatedRouteMock },
