@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AddCollectionComponent } from './add-collection.component';
 
@@ -6,7 +6,7 @@ describe('AddCollectionComponent', () => {
   let component: AddCollectionComponent;
   let fixture: ComponentFixture<AddCollectionComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AddCollectionComponent ]
     })
